+++ src/AltTextLab.php
@@ -24,7 +24,7 @@ class AltTextLab extends Plugin
     public static AltTextLab $plugin;
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '1.0.2';
+    public string $schemaVersion = '1.0.3';
 
     public function init(): void
     {